Create parseLIJobPage
I'm going to be honest, this Applescript text parser I wrote sucks, like it's terrible, but it's doing the trick to help me demonstrate value and see if I can get a prototype of my automated BDprinter idea working. Working well enough to see if it would actually be useful, that is. Input should be a plain text file of the text found on a LinkedIn job posting page.