Create getHyperlinks
Applescript to open the URL we created in userToURL and pull out every hyperlink on the page. I then run those hyperlinks through the "Filter URLs" command in the OSX Automator. I did not write this, I think I found it as a response on StackOverflow, and it worked with very minimal tweaking.