Create printSafariToPDF
Applescript to count tabs open in Safari, got this from another user on Github. Second part of Applescript is to print out each tab as a PDF. Sometimes this process begins to more more and more slowly when you run it. Sometimes it just blows right through printing all the pages to PDFs. For some reason I couldn't get the PDF printing section to work correctly with waiting to move to the next step until you confirmed the next window is open, so I just hacked it with delay statements.